feat(ChessmenInitializer): implement ChessmenInitializer domain
- move initializing chessmen logic to ChessmenInitializer
- change ChessGame constructor
